requirements.txt
Django==3.1.3
django-filter==2.4.0
django-graphql-jwt==0.3.1
PyJWT==1.7.1
graphene-django==2.13.0
psycopg2-binary==2.9.4
django-cors-headers==3.5.0
python-decouple
coverage
mixer